pyrt/requirements.txt
pytest==5.4.2
setuptools==46.4.0
numpy==1.18.1
Pillow==7.1.2
tqdm==4.46.0
redis==3.5.3
celery==4.4.0
requests==2.23.0
jsonpickle==1.4.1